requirements.txt
click
jsonschema
pandas
pyyaml
tabulate
wheel>=0.22
# lab sofrware
rna_seq_tools
rna_secstruct @ git+https://github.com/jyesselm/rna_secstruct@main
rna_map @ git+https://github.com/YesselmanLab/rna_map@main

